requirements.txt
langchain
flask
uvicorn
fastapi
openai
streamlit
langchain-core
langchain-community
langchain-openai
PyPDF2
faiss-cpu 
tiktoken 
unstructured
chromadb
python-dotenv
pinecone-client
cassio     
ctransformers
google-generativeai
langchain_google_genai
bitsandbytes
sentence-transformers
langserve
rapidocr-onnxruntime
sse_starlette 